chore: Move `database/` to `application/database/`
The source files in `database/` specify how the application manages
persisted data. It is conceptually subordinate to the notion
"application", which itself is conceptually separate from "models"
and "state".

The "models" have little to do with the data that is being persisted
and instead capture the blockchain protocol. One could conceive of
changes to the blockchain protocol that have little or no effect on
the persistence layer.

Likewise, the "state" module captures logic for synchronizing to the
rest of the network. This computational task makes abstraction of any
persistence requirements or functionality. Concretely, you could still
have a state that synchronizes to the blockchain without ever
persisting anything.

The reason why things are persisted to begin with, is so that the user
has continuity across runs of the application. It exists to benefit
the experience of users of the application.

This commit reflects this subordination to "application" in the
directory structure.
