+++ neptune-core/benchmark/benches/archival_mmr.rs
@@ -3,11 +3,11 @@ use leveldb::options::Options;
 use leveldb::options::ReadOptions;
 use leveldb::options::WriteOptions;
 use leveldb_sys::Compression;
-use neptune_cash::database::storage::storage_schema::traits::*;
-use neptune_cash::database::storage::storage_schema::DbtVec;
-use neptune_cash::database::storage::storage_schema::SimpleRustyStorage;
-use neptune_cash::database::storage::storage_vec::traits::StorageVecBase;
-use neptune_cash::database::NeptuneLevelDb;
+use neptune_cash::application::database::storage::storage_schema::traits::*;
+use neptune_cash::application::database::storage::storage_schema::DbtVec;
+use neptune_cash::application::database::storage::storage_schema::SimpleRustyStorage;
+use neptune_cash::application::database::storage::storage_vec::traits::StorageVecBase;
+use neptune_cash::application::database::NeptuneLevelDb;
 use neptune_cash::util_types::archival_mmr::ArchivalMmr;
 use rand::random;
 use tasm_lib::prelude::Digest;
