+++ neptune-core/src/application/database/storage/storage_schema/enums.rs
