+++ neptune-core/src/application/database/storage/storage_schema/schema.rs
@@ -41,7 +41,7 @@ use crate::locks::tokio::LockCallbackFn;
 /// ```
 /// # // note: compile_fail due to: https://github.com/rust-lang/rust/issues/67295
 /// # tokio_test::block_on(async {
-/// # use neptune_cash::database::storage::{storage_vec::traits::*, storage_schema::{SimpleRustyStorage, traits::*}};
+/// # use neptune_cash::application::database::storage::{storage_vec::traits::*, storage_schema::{SimpleRustyStorage, traits::*}};
 /// # let db = neptune_cash::database::NeptuneLevelDb::open_new_test_database(true, None, None, None).await.unwrap();
 /// use neptune_cash::locks::tokio::AtomicRw;
 ///
