+++ neptune-core/src/application/database/storage/storage_schema/simple_rusty_storage.rs
@@ -5,7 +5,7 @@ use super::RustyKey;
 use super::RustyValue;
 use super::SimpleRustyReader;
 use super::WriteOperation;
-use crate::database::neptune_leveldb::WriteBatchAsync;
+use crate::application::database::neptune_leveldb::WriteBatchAsync;
 use crate::locks::tokio::LockCallbackFn;
 
 /// Database schema and tables logic for RustyLevelDB.
