+++ neptune-core/src/application/database/storage/storage_vec/traits.rs
@@ -1,7 +1,7 @@
 //! Traits that define the StorageVec interface
 //!
 //! It is recommended to wildcard import these with
-//! `use crate::database::storage::storage_vec::traits::*`
+//! `use crate::application::database::storage::storage_vec::traits::*`
 
 // for Stream (async Iterator equiv)
 use async_stream::stream;
@@ -123,7 +123,7 @@ pub trait StorageVecStream<T: Send>: StorageVecBase<T> {
     /// # Example:
     /// ```
     /// # tokio_test::block_on(async {
-    /// # use neptune_cash::database::storage::storage_vec::{OrdinaryVec, traits::*};
+    /// # use neptune_cash::application::database::storage::storage_vec::{OrdinaryVec, traits::*};
     /// # let mut vec = OrdinaryVec::<u32>::from(vec![1,2,3,4,5,6,7,8,9]);
     ///
     /// let stream = vec.stream().await;
@@ -147,7 +147,7 @@ pub trait StorageVecStream<T: Send>: StorageVecBase<T> {
     /// # Example:
     /// ```
     /// # tokio_test::block_on(async {
-    /// # use neptune_cash::database::storage::storage_vec::{OrdinaryVec, traits::*};
+    /// # use neptune_cash::application::database::storage::storage_vec::{OrdinaryVec, traits::*};
     /// # let mut vec = OrdinaryVec::<u32>::from(vec![1,2,3,4,5,6,7,8,9]);
     ///
     /// let stream = vec.stream_values().await;
@@ -171,7 +171,7 @@ pub trait StorageVecStream<T: Send>: StorageVecBase<T> {
     /// # Example:
     /// ```
     /// # tokio_test::block_on(async {
-    /// # use neptune_cash::database::storage::storage_vec::{OrdinaryVec, traits::*};
+    /// # use neptune_cash::application::database::storage::storage_vec::{OrdinaryVec, traits::*};
     /// # let mut vec = OrdinaryVec::<u32>::from(vec![1,2,3,4,5,6,7,8,9]);
     ///
     /// let stream = vec.stream_many([2,3,7]);
@@ -201,7 +201,7 @@ pub trait StorageVecStream<T: Send>: StorageVecBase<T> {
     /// # Example:
     /// ```
     /// # tokio_test::block_on(async {
-    /// # use neptune_cash::database::storage::storage_vec::{OrdinaryVec, traits::*};
+    /// # use neptune_cash::application::database::storage::storage_vec::{OrdinaryVec, traits::*};
     /// # let mut vec = OrdinaryVec::<u32>::from(vec![1,2,3,4,5,6,7,8,9]);
     ///
     /// let stream = vec.stream_many_values([2,3,7]);
