+++ neptune-core/src/application/mod.rs
@@ -1,1 +1,2 @@
 pub mod config;
+pub mod database;
