+++ neptune-core/src/state/archival_state.rs
@@ -20,10 +20,10 @@ use super::shared::new_block_file_is_needed;
 use super::StorageVecBase;
 use crate::api::export::Network;
 use crate::application::config::data_directory::DataDirectory;
-use crate::database::create_db_if_missing;
-use crate::database::storage::storage_schema::traits::*;
-use crate::database::NeptuneLevelDb;
-use crate::database::WriteBatchAsync;
+use crate::application::database::create_db_if_missing;
+use crate::application::database::storage::storage_schema::traits::*;
+use crate::application::database::NeptuneLevelDb;
+use crate::application::database::WriteBatchAsync;
 use crate::models::blockchain::block::block_header::BlockHeader;
 use crate::models::blockchain::block::block_header::BlockHeaderWithBlockHashWitness;
 use crate::models::blockchain::block::block_header::HeaderToBlockHashWitness;
@@ -1419,7 +1419,7 @@ pub(super) mod tests {
     use crate::application::config::cli_args;
     use crate::application::config::data_directory::DataDirectory;
     use crate::application::config::network::Network;
-    use crate::database::storage::storage_vec::traits::*;
+    use crate::application::database::storage::storage_vec::traits::*;
     use crate::mine_loop::tests::make_coinbase_transaction_from_state;
     use crate::models::blockchain::block::block_transaction::BlockTransaction;
     use crate::models::blockchain::consensus_rule_set::ConsensusRuleSet;
