+++ neptune-core/src/state/mod.rs
@@ -62,9 +62,9 @@ use crate::api::export::NeptuneProof;
 use crate::application::config::cli_args;
 use crate::application::config::data_directory::DataDirectory;
 use crate::application::config::tx_upgrade_filter::TxUpgradeFilter;
-use crate::database::storage::storage_schema::traits::StorageWriter as SW;
-use crate::database::storage::storage_vec::traits::*;
-use crate::database::storage::storage_vec::Index;
+use crate::application::database::storage::storage_schema::traits::StorageWriter as SW;
+use crate::application::database::storage::storage_vec::traits::*;
+use crate::application::database::storage::storage_vec::Index;
 use crate::locks::tokio as sync_tokio;
 use crate::locks::tokio::AtomicRwReadGuard;
 use crate::locks::tokio::AtomicRwWriteGuard;
