+++ neptune-core/src/state/wallet/wallet_state.rs
@@ -47,12 +47,12 @@ use super::wallet_status::WalletStatusElement;
 use crate::application::config::cli_args::Args;
 use crate::application::config::data_directory::DataDirectory;
 use crate::application::config::fee_notification_policy::FeeNotificationPolicy;
-use crate::database::storage::storage_schema::DbtVec;
-use crate::database::storage::storage_schema::RustyKey;
-use crate::database::storage::storage_schema::RustyValue;
-use crate::database::storage::storage_vec::traits::*;
-use crate::database::storage::storage_vec::Index;
-use crate::database::NeptuneLevelDb;
+use crate::application::database::storage::storage_schema::DbtVec;
+use crate::application::database::storage::storage_schema::RustyKey;
+use crate::application::database::storage::storage_schema::RustyValue;
+use crate::application::database::storage::storage_vec::traits::*;
+use crate::application::database::storage::storage_vec::Index;
+use crate::application::database::NeptuneLevelDb;
 use crate::mine_loop::composer_outputs;
 use crate::mine_loop::composer_parameters::ComposerParameters;
 use crate::models::blockchain::block::block_height::BlockHeight;
@@ -426,7 +426,7 @@ impl WalletState {
     }
 
     async fn open_wallet_db(path: &Path) -> anyhow::Result<NeptuneLevelDb<RustyKey, RustyValue>> {
-        NeptuneLevelDb::new(path, &crate::database::create_db_if_missing())
+        NeptuneLevelDb::new(path, &crate::application::database::create_db_if_missing())
             .await
             .map_err(|e| anyhow::anyhow!("Failed to open wallet db at '{}': {}", path.display(), e))
     }
@@ -3743,7 +3743,7 @@ pub(crate) mod tests {
 
         mod worker {
             use super::*;
-            use crate::database::storage::storage_schema::traits::StorageWriter;
+            use crate::application::database::storage::storage_schema::traits::StorageWriter;
             use crate::tests::shared::files::unit_test_data_directory;
 
             /// tests that all known keys are unique for a given key-type
@@ -3996,7 +3996,7 @@ pub(crate) mod tests {
 
         mod worker {
             use super::*;
-            use crate::database::storage::storage_schema::traits::StorageWriter;
+            use crate::application::database::storage::storage_schema::traits::StorageWriter;
             use crate::tests::shared::files::unit_test_data_directory;
 
             /// implements a test with 2 variations via `persist` param.
