+++ neptune-core/src/util_types/archival_mmr.rs
@@ -14,8 +14,8 @@ use tasm_lib::twenty_first::util_types::mmr::shared_advanced::node_index_to_leaf
 use tasm_lib::twenty_first::util_types::mmr::shared_basic::leaf_index_to_mt_index_and_peak_index;
 use tasm_lib::twenty_first::util_types::mmr::shared_basic::right_lineage_length_from_leaf_index;
 
-use crate::database::storage::storage_schema::DbtVec;
-use crate::database::storage::storage_vec::traits::*;
+use crate::application::database::storage::storage_schema::DbtVec;
+use crate::application::database::storage::storage_vec::traits::*;
 use crate::models::blockchain::shared::Hash;
 
 /// A Merkle Mountain Range is a datastructure for storing a list of hashes.
@@ -392,10 +392,10 @@ pub(crate) mod tests {
     use test_strategy::proptest;
 
     use super::*;
-    use crate::database::storage::storage_schema::traits::*;
-    use crate::database::storage::storage_schema::SimpleRustyStorage;
-    use crate::database::storage::storage_vec::OrdinaryVec;
-    use crate::database::NeptuneLevelDb;
+    use crate::application::database::storage::storage_schema::traits::*;
+    use crate::application::database::storage::storage_schema::SimpleRustyStorage;
+    use crate::application::database::storage::storage_vec::OrdinaryVec;
+    use crate::application::database::NeptuneLevelDb;
     use crate::tests::shared_tokio_runtime;
     use crate::Hash;
