+++ neptune-core/src/util_types/mutator_set/rusty_archival_mutator_set.rs
@@ -3,13 +3,13 @@ use tasm_lib::twenty_first::tip5::digest::Digest;
 use super::active_window::ActiveWindow;
 use super::archival_mutator_set::ArchivalMutatorSet;
 use super::removal_record::chunk::Chunk;
-use crate::database::storage::storage_schema::traits::*;
-use crate::database::storage::storage_schema::DbtSingleton;
-use crate::database::storage::storage_schema::DbtVec;
-use crate::database::storage::storage_schema::RustyKey;
-use crate::database::storage::storage_schema::RustyValue;
-use crate::database::storage::storage_schema::SimpleRustyStorage;
-use crate::database::NeptuneLevelDb;
+use crate::application::database::storage::storage_schema::traits::*;
+use crate::application::database::storage::storage_schema::DbtSingleton;
+use crate::application::database::storage::storage_schema::DbtVec;
+use crate::application::database::storage::storage_schema::RustyKey;
+use crate::application::database::storage::storage_schema::RustyValue;
+use crate::application::database::storage::storage_schema::SimpleRustyStorage;
+use crate::application::database::NeptuneLevelDb;
 use crate::util_types::archival_mmr::ArchivalMmr;
 
 type AmsMmrStorage = DbtVec<Digest>;
